Analytics | Store baseUrl on the API Request (#294)
Add base url as a field on the API request so it can be stored in the queue, and overridden in subclasses.
Incidentally, fixes an old but trivial bug where switching base URLs (i.e. local/ngrok/prod) in the config could cause queued requests to wind up sending to the wrong server.

Co-authored-by: Evan Masseau <>