Sample App (#289)
Adds a basic sample app to the repo to use for integration demo, and serve as an in-repo application for manual and automated testing.

I've also done a bit of workflow adjustments and clean up in here, no significant changes in the requirements or core functionality, mostly quality of life stuff, prompted by the sample app addition.

---------

Co-authored-by: Evan Masseau <>